Removed extra line feeds, standardized Given_..Then_ of tests method names. Also moved DateTimeOffsetModifier to own class.
